feat: Add font & Code Refactoring
