+++ src/app.css
@@ -1,3 +1,5 @@
+@import url(//spoqa.github.io/spoqa-han-sans/css/SpoqaHanSansNeo.css);
+
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
@@ -20,6 +22,9 @@
 	border: 1px solid #e8e8e8;
 }
 
+* {
+	font-family: 'Spoqa Han Sans Neo', 'sans-serif';
+}
 html,
 body,
 #app {
